refactor(expr): split to concerns and contracts (#22)
