+++ README.md
@@ -215,10 +215,10 @@ php bin/structura analyze
   - [toBeAttribute()](#tobeattribute)
 - 🔗 Dependencies
   - [dependsOnlyOn()](#dependsonlyon)
-  - [dependsOnlyOnAttribut](#dependsonlyonattribut)
-  - [dependsOnlyOnImplementation](#dependsonlyonimplementation)
-  - [dependsOnlyOnInheritance](#dependsonlyoninheritance)
-  - [dependsOnlyOnUseTrait](#dependsonlyonusetrait)
+  - [dependsOnlyOnAttribut()](#dependsonlyonattribut)
+  - [dependsOnlyOnImplementation()](#dependsonlyonimplementation)
+  - [dependsOnlyOnInheritance()](#dependsonlyoninheritance)
+  - [dependsOnlyOnUseTrait()](#dependsonlyonusetrait)
   - [toNotDependsOn()](#tonotdependson)
   - [dependsOnFunction()](#dependsonfunction)
   - [toNotDependsOnFunction()](#tonotdependsonfunction)
