+++ tests/Unit/Services/AnalyseServiceTest.php
@@ -66,7 +66,7 @@ final class AnalyseServiceTest extends TestCase
              & to extend <promote>BadMethodCallException</promote>
 
         <pass> PASS </pass> Asserts architecture rules in StructuraPhp\Structura\Tests\Feature\TestVoid
-        78 classes from
+        90 classes from
          - dirs
         That
         Should
